// backend/Audit-Log/Model/Logs.ts
export enum Logs {

    EMAIL_VERIFIED = 'EMAIL_VERIFIED',
    USER_REGISTERED = 'USER_REGISTERED',
    USER_CHANGED_EMAIL = 'USER_CHANGED_EMAIL',
    OTP_SENT = 'OTP_SENT',
    OTP_SEND_FAILED = 'OTP_SEND_FAILED',

    LOGIN_SUCCESS = 'LOGIN_SUCCESS',
    LOGIN_FAILED = 'LOGIN_FAILED',
    LOGOUT = 'LOGOUT',

    PASSWORD_RESET_REQUESTED = 'PASSWORD_RESET_REQUESTED',
    PASSWORD_RESET_COMPLETED = 'PASSWORD_RESET_COMPLETED',

    MFA_ENABLED = 'MFA_ENABLED',
    MFA_DISABLED = 'MFA_DISABLED',
    PASSWORD_CHANGED = 'PASSWORD_CHANGED',
    PROFILE_UPDATED = 'PROFILE_UPDATED',

    UNAUTHORIZED_ACCESS = 'UNAUTHORIZED_ACCESS',
    TOKEN_BLACKLISTED = 'TOKEN_BLACKLISTED',
    RBAC_DENIED = 'RBAC_DENIED',
    RBAC_AUTHORIZED = 'RBAC_AUTHORIZED',

    ADMIN_CREATED_USER = 'ADMIN_CREATED_USER',
    ADMIN_ANNOUNCE_ALL = 'ADMIN_ANNOUNCE_ALL',
    ROLE_CHANGED = 'ROLE_CHANGED',
    ADMIN_ANNOUNCE_ROLE = 'ADMIN_ANNOUNCE_ROLE',
    ADMIN_DELETE_USER = 'ADMIN_DELETE_USER',
    DATA_EXPORT = 'DATA_EXPORT',

    DATA_BACKUP_STARTED = 'DATA_BACKUP_STARTED',
    DATA_BACKUP_COMPLETED = 'DATA_BACKUP_COMPLETED',
    DATA_BACKUP_FAILED = 'DATA_BACKUP_FAILED',

    DATA_BACKUP_TEST_STARTED = 'DATA_BACKUP_TEST_STARTED',
    DATA_BACKUP_TEST_COMPLETED = 'DATA_BACKUP_TEST_COMPLETED',
    DATA_BACKUP_TEST_FAILED = 'DATA_BACKUP_TEST_FAILED',

    //CHAT CREATED, FEEDBACK CREATED, SENT, READ ADD MORE ETC...
    ADMIN_FORCE_LOGOUT='ADMIN_FORCE_LOGOUT',
    ADMIN_USER_LOCK='ADMIN_USER_LOCK',
    ADMIN_USER_UNLOCK='ADMIN_USER_UNLOCK',

}